Fixed SensorDefinition type errors.
